requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
unimol-tools>=1.0.0
PyYAML>=6.0
psutil>=5.8.0

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional performance dependencies
numba>=0.56.0
joblib>=1.2.0